FIX :: STEP2 리팩토링 (테스트명 수정/ describe 활용, 상황과 동작 분리)
