--- __tests__/domain/Car/Car.test.const.js
@@ -1,1 +1,0 @@
-export const DEFAULT_CAR_NAME = "sonata";
