+++ src/domain/Car/Car.test.const.js
@@ -1,0 +1,1 @@
+export const DEFAULT_CAR_NAME = "ray";
